refactor(files): OS へパスを渡す 2 ルートの起動ループを 1 つにする（セルフレビュー）
マージ前に差分を通しで読んで、自分で 2 点見つけた。

## 1. `open-file.ts` が `reveal.ts` とほぼ同型だった

候補リストの巡回、WSL のパス変換、「起動したら答える」規則、全試行を並べる 500 —
**同じ推論の写しが 2 つ**あった。違うのは argv だけ。プロセス起動のコードで写しが 2 つあると、
片方だけ直されて drift する（WSL のフォールバックはまさにそれを防ぐために入ったもの）。

`spawnFirstOpener(target, argvFor, spawner)` に括り出した。argv の組み立てだけを関数で
受け取るので、**2 ルートが違うのは argv だけ**であることがコードの形に出る。

  reveal.ts    67 → 44 行
  open-file.ts 63 → 37 行
  spawnOpener.ts  50 行（新規）

挙動不変の証拠: 両ルートの spec 16 件がそのまま通る。反転 mutation 2 種も赤くなる
（2 ルートの argv を取り違える / 何も起動しなくても ok を返す）。実サーバでも確認:

  open   (ファイル) → open <ws>/book.xlsx
  reveal (ファイル) → open -R <ws>/book.xlsx
  reveal (フォルダ) → open <ws>
  open   (フォルダ) → 400（アプリを起動するルートにフォルダを渡さない）

## 2. `contentTypeKey` が外から使われないのに export されていた

非公開にした。knip は通っていたが、公開面は小さいほうがよい。
